+++ utilities/mixins/openapi/subscription.openapi.mixin.ts
@@ -313,6 +313,7 @@ const SubscriptionCreateOpenapi = {
             coupon: { type: 'string' },
             grantTo: { type: 'string', format: 'url', description: 'This field is used to donor the subscription to another store' },
             autoRenew: { type: 'boolean' },
+            dueDate: { type: 'string', required: false, format: 'date', example: 'yyyy-mm-dd' },
           },
         },
       },